Full Project/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
faiss-cpu
numpy
deep-translator
requests #==2.31.0
ibm-watsonx-ai #==1.0.0

